send ig

+++ Model/Location.cs
@@ -2,6 +2,6 @@ namespace ClassLibraryDAL;
 
 public class Location
 {
-    public float Long { get; set; } = 0;
-    public float Lat { get; set; } = 0;
+    public double Long { get; set; } = 0;
+    public double Lat { get; set; } = 0;
 }
